Various optimization fixes for basic tests
Finalized updates to bash scripts

Fixed error in config files that was using a wrong value for the history
table memory limit, of 0.5 instead of 0.8, holdover from previous
experimentation that wasn't originally caught.

Added edge cases to retrieve_input to prevent it breaking on tsplib
instances due to different formatting.